+++ src/main/java/com/esliceu/movies/services/GenderServices.java
@@ -16,7 +16,8 @@ import java.util.stream.Collectors;
 public class GenderServices {
     @Autowired
     GenderRepo genderRepo;
-
+    @Autowired
+    PermissionsServices permissionsServices;
 
     public Page<Gender> findAllGenders(Pageable pageable) {
         return genderRepo.findAll(pageable);
@@ -27,23 +28,9 @@ public class GenderServices {
         List<String> names = genders.stream()
                 .map(g -> g.getGender())
                 .collect(Collectors.toList());
-
         Gson gson = new Gson();
         String result = gson.toJson(names);
         return result;
-    }
-
-    public String saveGender(String name) {
-        if (name == null || name.trim().isEmpty()) {
-            return "El genero no puede estar vacío.";
-        }
-        if (genderRepo.findGenderByGender(name).size() >= 1) {
-            return "Ya existe un genero con ese nombre.";
-        }
-        Gender gender = new Gender();
-        gender.setGender(name);
-        genderRepo.save(gender);
-        return null;
     }
 
     public Gender findGenderById(Integer id) {
@@ -54,26 +41,36 @@ public class GenderServices {
         return genderRepo.findGenderByGender(name);
     }
 
-    public String deleteGender(Integer id) {
+    public String saveGender(String name, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Crear géneros");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
+        if (name == null || name.trim().isEmpty()) return "El genero no puede estar vacío.";
+        if (genderRepo.findGenderByGender(name).size() >= 1) return "Ya existe un genero con ese nombre.";
+        Gender gender = new Gender();
+        gender.setGender(name);
+        genderRepo.save(gender);
+        return null;
+    }
+
+    public String deleteGender(Integer id, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Eliminar géneros");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
         try {
             genderRepo.deleteById(id);
-            return "Ok";
-        }catch (Exception e) {
-            return "Error";
-        }
-
-    }
-
-    public Gender updateGender(Integer id, String name) {
-        Optional<Gender> existingGender = genderRepo.findById(id);
-        if (existingGender.isPresent()) {
-           Gender updatedGender = existingGender.get();
-            updatedGender.setGender(name);
-            genderRepo.save(updatedGender);
-            return updatedGender;
-        } else {
             return null;
+        } catch (Exception e) {
+            return "Ha habido un error al eliminar el género";
         }
     }
 
+    public String updateGender(Integer id, String name, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Modificar géneros");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
+        Optional<Gender> existingGender = genderRepo.findById(id);
+        if (existingGender.isEmpty()) return "No existe ese género";
+        Gender updatedGender = existingGender.get();
+        updatedGender.setGender(name);
+        genderRepo.save(updatedGender);
+        return null;
+    }
 }
