+++ src/main/java/com/esliceu/movies/services/KeywordServices.java
@@ -18,6 +18,9 @@ public class KeywordServices {
     @Autowired
     KeywordRepo keywordRepo;
 
+    @Autowired
+    PermissionsServices permissionsServices;
+
     public Page<Keyword> findAllKeywords(Pageable pageable) {
         return keywordRepo.findAll(pageable);
     }
@@ -33,20 +36,6 @@ public class KeywordServices {
         return result;
     }
 
-    public String saveKeyword(String name) {
-        if (name == null || name.trim().isEmpty()) {
-            return "El nombre de la keyword no puede estar vacío.";
-        }
-        if (keywordRepo.findKeywordByKeywordName(name).size() >= 1) {
-            return "Ya existe una keyword con ese nombre.";
-        }
-        Keyword keyword = new Keyword();
-        keyword.setKeywordName(name);
-        keywordRepo.save(keyword);
-        return null;
-    }
-
-
     public Keyword findKeywordById(Integer id) {
         return keywordRepo.findById(id).get();
     }
@@ -55,26 +44,37 @@ public class KeywordServices {
         return keywordRepo.findKeywordByKeywordName(keywordSearch);
     }
 
-    public String deleteKeyword(Integer id) {
+    public String saveKeyword(String name, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Crear palabras clave");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
+        if (name == null || name.trim().isEmpty()) return "El nombre de la keyword no puede estar vacío.";
+        if (keywordRepo.findKeywordByKeywordName(name).size() >= 1) return "Ya existe una keyword con ese nombre.";
+        Keyword keyword = new Keyword();
+        keyword.setKeywordName(name);
+        keywordRepo.save(keyword);
+        return null;
+    }
+
+    public String deleteKeyword(Integer id, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Eliminar palabras clave");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
         try {
             keywordRepo.deleteById(id);
-            return "Ok";
-        }catch (Exception e) {
-            return "Error";
-        }
-
-    }
-
-    public Keyword updateKeyword(Integer id, String name) {
-        Optional<Keyword> existingKeyword = keywordRepo.findById(id);
-        if (existingKeyword.isPresent()) {
-            Keyword updatedKeyword = existingKeyword.get();
-            updatedKeyword.setKeywordName(name);
-            keywordRepo.save(updatedKeyword);
-            return updatedKeyword;
-        } else {
             return null;
+        } catch (Exception e) {
+            return "Ha habido un error al eliminar la palabra clave";
         }
+
     }
 
+    public String updateKeyword(Integer id, String name, String username) {
+        String necessaryPermission = permissionsServices.checkPermisions(username, "Modificar palabras clave");
+        if (necessaryPermission == null) return "No tienes el permiso necesario";
+        Optional<Keyword> existingKeyword = keywordRepo.findById(id);
+        if (existingKeyword.isEmpty()) return "No existe esa palabra clave";
+        Keyword updatedKeyword = existingKeyword.get();
+        updatedKeyword.setKeywordName(name);
+        keywordRepo.save(updatedKeyword);
+        return null;
+    }
 }
